Move tests directory into src
